Add custom route to create custom lists.
Use express routing parameters to create a custom list route,
that checks if the route exists using findOne() method. It creates a
new lists if it doesnt exist, and renders the list if it already exists
in the lists database. Create a new Schema for the lists and make its
model to call in the custome list route.